Move class to Blocks folder.
