+++ app/di.php
@@ -7,11 +7,12 @@ $registrar->addInstance(new \Dxw\Iguana\Theme\LayoutRegister(
 ));
 
 // Blocks
+$registrar->addInstance(new \GovUKBlogs\Blocks\BlockCategory());
 $registrar->addInstance(new \GovUKBlogs\Blocks\Details\Block());
 $registrar->addInstance(new \GovUKBlogs\Blocks\InsetText\Block());
 
 // Theme
-$registrar->addInstance(new \GovUKBlogs\Theme\BlockCategory());
+
 $registrar->addInstance(new \GovUKBlogs\Theme\ThemeSupports());
 $registrar->addInstance(new \GovUKBlogs\WidgetCategoriesDropdownWithSubmit\Register());
 $registrar->addInstance(new \GovUKBlogs\WidgetArchiveDropdownWithSubmit\Register());
